move common testing functions to testing package
